+++ IGUIDE/Raster.h
@@ -10,11 +10,11 @@ class Raster
 public:
 	vector<CD2DPointF>	corner;								// raster corners from subject view
 	vector<Edge>		perimeter;							// raster perimeter
-	float				size;								// raster size
+	double				size;								// raster size
 	D2D1_COLOR_F		color;								// raster color
 	CD2DPointF			scale;								// for resizing operator's view
-	float				meanAlpha;							// displacement angle
-	float				meanEdge;							// mean edge
+	double				meanAlpha;							// displacement angle
+	double				meanEdge;							// mean edge
 	CD2DPointF			mid;								// triangulate mid point of raster
 
 	Raster();
